+++ classes/class-wc-gateway-klarna-payments.php
@@ -11,7 +11,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 
 use Krokedil\Klarna\Features;
 use Krokedil\Klarna\PluginFeatures;
-use Krokedil\Klarna\KlarnaExpressCheckout\KECOneStepIntegration;
+use Krokedil\Klarna\ExpressCheckout\KECOneStepIntegration;
 use KrokedilKlarnaPaymentsDeps\Krokedil\SettingsPage\SettingsPage;
 
 /**
@@ -119,7 +119,7 @@ class WC_Gateway_Klarna_Payments extends WC_Payment_Gateway {
 		$this->hide_what_is_klarna  = 'yes' === $this->get_option( 'hide_what_is_klarna' );
 		$this->float_what_is_klarna = 'yes' === $this->get_option( 'float_what_is_klarna' );
 
-		$this->pay_button_id = Krokedil\Klarna\KlarnaExpressCheckout::get_payment_button_id();
+		$this->pay_button_id = Krokedil\Klarna\ExpressCheckout::get_payment_button_id();
 
 		// Hooks.
 		add_action( 'woocommerce_update_options_payment_gateways_' . $this->id, array( $this, 'process_admin_options' ) );
@@ -411,7 +411,7 @@ class WC_Gateway_Klarna_Payments extends WC_Payment_Gateway {
 
 		if ( 'redirect' === $checkout_flow ) {
 			// If the user arrived via the KEC two-step flow, skip the HPP and process the order directly.
-			$kec_client_token = Krokedil\Klarna\KlarnaExpressCheckout\Session::get_client_token();
+			$kec_client_token = Krokedil\Klarna\ExpressCheckout\Session::get_client_token();
 			if ( empty( $kec_client_token ) ) {
 				return $this->process_blocks_order( $order );
 			}
@@ -439,7 +439,7 @@ class WC_Gateway_Klarna_Payments extends WC_Payment_Gateway {
 	 * @return array
 	 */
 	private function process_checkout_order( $order ) {
-		$kec_client_token = Krokedil\Klarna\KlarnaExpressCheckout\Session::get_client_token();
+		$kec_client_token = Krokedil\Klarna\ExpressCheckout\Session::get_client_token();
 		$order_key        = $order->get_order_key();
 		$order_id         = $order->get_id();
