+++ src/ExpressCheckout/assets/css/kec-admin.css
