+++ src/ExpressCheckout/assets/css/kec-cart.css
