+++ src/ExpressCheckout/assets/img/preview-default-dark.png
