+++ src/ExpressCheckout/assets/img/preview-default-light.png
