+++ src/ExpressCheckout/assets/img/preview-default-outlined.png
