+++ src/ExpressCheckout/assets/img/preview-pill-dark.png
