+++ src/ExpressCheckout/assets/img/preview-pill-light.png
