+++ src/ExpressCheckout/assets/img/preview-pill-outlined.png
