+++ src/ExpressCheckout/assets/img/preview-rect-dark.png
