+++ src/ExpressCheckout/assets/img/preview-rect-light.png
