+++ src/ExpressCheckout/assets/img/preview-rect-outlined.png
